# [exporter/loadbalancing] Improve the performance when merging traces belonging to the same backend