--- a/.chloggen/loadbalancingexporter_perf.yaml
+++ b/.chloggen/loadbalancingexporter_perf.yaml
@@ -7,7 +7,7 @@ change_type: enhancement
 component: loadbalancingexporter
 
 # A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
-note: Improve performance
+note: Improve the performance when merging traces belonging to the same backend 
 
 # Mandatory: One or more tracking issues related to the change. You can use the PR number here if no issue exists.
 issues: [32032]