@lauri-paypay 2024-04-16
@dmitryax sorry for mentioning you directly, but since you seem to have been active in maintaining loadbalancingexporter, could I ask you to take a look at this PR?