@jpkrohling 2024-04-30
@lauri-paypay, thank you for the PR. The problem that PR was solving is optimizing the number of requests made by the load balancing exporter to the next layer of collectors. By merging traces per endpoint, we are making traces that are meant to the same destination to be sent on a single payload. 

Your benchmark here is good, but it focuses only on a specific function call. Would it be possible to also compare the number of requests (or calls to `consumeMetric`)? Or even better, perhaps we could also report the number of entries on the payload as part of one of the available metrics?