@lauri-paypay 2024-05-01
>Your benchmark here is good, but it focuses only on a specific function call. Would it be possible to also compare the number of requests (or calls to consumeMetric)? Or even better, perhaps we could also report the number of entries on the payload as part of one of the available metrics?

I'm sure it's possible, but I don't think that should be done as part of this PR, because this PR changes only the function that I did test.