@lauri-paypay 2024-05-08
> I still feel like we are missing something, but I'm OK giving this a go, as we do have test coverage for the scenario I have in mind.
> 
> The only request I have: can we make the changelog entry a bit easier for users to find out what's happening. In case of problems, they should be able to come back to this PR and let us know they are having problems.

thanks. I committed your suggestion to address that.